# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "My first swagger documented API",
        "version": "1.0.0"
    },
    "paths": {
        "/blx/public/api/v1/login": {
            "post": {
                "summary": "Check the login credentials and get the access token",
                "description": "Check the login credentials and get the access token",
                "operationId": "login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "description": "User email",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "path",
                        "description": "User password",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/blx/public/api/v1/users": {
            "get": {
                "summary": "Get all users",
                "description": "Get all users",
                "operationId": "users",
                "parameters": [
                    {
                        "name": "auth",
                        "in": "header",
                        "description": "an authorization header",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        }
    },
    "definitions": {}
}